feedback display for each course by id is implemented for admin.viewing of list of courses for admin is changed. Admin can view each course in detail including feedbacks for the course by course id

+++ BackendModule/src/main/java/com/LDMSAppBackend/BackendModule/repositories/CourseAssignmentRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CourseAssignmentRepository extends JpaRepository<CourseAssignment, Integer> {
+public interface CourseAssignmentRepository extends JpaRepository<CourseAssignment, Long> {
     List<CourseAssignment> findByEmployee_EmployeeId(Integer employeeId);
     Optional<CourseAssignment> findByCourse_CourseIdAndEmployee_EmployeeId(Long courseId, Integer employeeId);
     Boolean existsByEmployee_EmployeeIdAndCourse_CourseId(Integer employeeId, Long courseId);
